fix: final bugs. AI now works 100%
Remaining bugs:
- in getNextMove the AI player wasn't actually placing
the move on the board before evaluating the new board, effectively
evaluating the same board every time.
- the BoardEvaluator judged the boards wrongly: instead of giving a score
depending on who won, it gave different scores depending on who the
current player was (max/min) and if their symbol was winning.
- the "optimization" that added/subtracted the depth was also operating
on the score made when the board indicated a draw.