feat: allow starting a new game from UI
Also refactored the the way players do their move, and how switching
between players is handled, both in the controller and the game itself.
The (wrong) idea implemented before was that the starting symbol changes,
but actually X always starts, and it's the players that get assigned
a different symbol.